chore(release): creates auto signoff script
the script is setup in such a way to rm the need for any user prompts. If successful it also selfdestructs/rm itself. By wrapping EOF tag in single quotes, we avoid execution of any of the shell parameters within the string, without it they execute and could lead to unwanted removal of the release script etc...