feat: clear child pass on other inputs changing
feature based on user request. Requested to help issues where a user may have changed options or inputs, but forgot to click generate, and then pastes a previously generated password. This avoids that, as whatever the current inputs are set to, can be applicable only to the shown child pass, or if none is shown, it means it must be genereated. Added triggers and baseInputs/advInputs vars for more flexible management of inputs if the ffuture need arises, through reactivity.