feat: add SKIP option to verify step
By typing SKIP, verification can be ended early, as some users complained it is too onerous, especially when importing a previous backup. However, a note is left indicating they are breaking their warranty, i.e. that it is up to them to have then ensured and verified their phrase is saved correctly.