chore: add babel to build for es2020 support
since we are using webpack 4, it utilizes an older version of acorn, 6 to be specific, while 8 is needed for es2020 features to work. We anyways anticipated using babel... and now we're just actually enabling it in the build process, we needed it anyways for the tests to work. Bundle sizes did not significantly differ after the inclusion.