feat: provide users with 3 improved setup options
this commit still requires some changes for the quick setup to be a valid step. It requires the plain backup phrase to be kept in the browser, until the user verifies from within the app at a later time. They should get prompted about this.

Many changes to the NavModal to accomodate this, and if anything it has improved it vastly and made it even more generalized. Specifically by foregoing the previous interaction of components via svelte stores, now we just pass props fore and aft parents/children as necessary.

The setup process now has a selector and processor. Both should be clear
in what they actually do.

The firstVisit modal paths have been kept as is for best git diffing,
however, will be better divided up and maybe even further, with possibly
specific finish messages depending on setup.

The textual content of each modal has been generally kept as is... and
to be changed later.

ImportMnemonic is an exception to this, as we changed the importation
from an optionality, to ofc a necessary step here, when being dedicated
to just the existing user setup flow.