feat: save & encrypt backup phrase if necessary
for delayed verification during app usage, which is a TODO feature

backup phrase === mnemonic in the code, it is more succinct and
standard, however, backup phrase was decided in user-facing context to
make it cleare to end-users, they should make a backup of it.

CHANGES

extend mnemonic store to support verify and skipped properties for the new setup flow
adds plainBlob implementation, whose creation is triggered in certain
instances
added necessary methods for the plain mnemonic encryption and this in turn resulted in some of the codebase there becoming more generalized, so the seed encryption could be re-used
hooked verified being appropriately set from existing mnemonic store
functions where possible
elsewhere hooked the appropriate verify and skip calls across the
necessary modals

RATIONALE

this feature was done due to the new 3 setup flow

full
quick
existing

for existing users, the saving of the backup phrase in the browser is
completely unnecessary, as they have proven they have it backuped in the
first place, to be able to import it

in the case of quick users... the mnemonic is never shown to them in a
drive to get them using the app as fast as possible and not overencumber
them with the setup steps. Therefore, this feature was ultimately
required, as it allows a delay of the mnemonic phrase feedback and
verification during app use (TODO feature)

for full setup users, verifying ALL 12 words, it is also not required.
However, in case a full user SKIPs verification, the backup phrase will
also be saved.

UPCOMING

For users that have the plainBlob, they will be prompted to verify their
mnemonic

The flow will ask them, if they have their backup phrase already or not

if they do not, they will be prompted for their pass, to decrypt the
blob and give them an opportunity to back it up

thereafter they will be provided with an opportunity to fully verify
their mnemonic.

upon success of that... the plainBlob is deleted

as can be seen, in the case of full skipped users that already made a
backup, the plainBlob may not even get decrypted once during its
lifetime.