feat: add improved mobile tooltips globally
enables mobile tooltips globally, with an imo vast UX improvement compared to how they were, whereby with pure CSS we just fade it out after a while, and also use a positioining/transform trick to get a similar effect to native tooltips, where the tooltip can outsize its parent. Additionally adds ? to indicate there is a tooltip... however may remove this, as this feature was based on long-clicking to get a tooltip out, however the UX of that turned out it sucked.