fix: properly overwrite mnemonic for post verify backup phrase
a bug was found which could occur in certain instances with the quick setup or skipped full setup flows. To recrate it prior to this commit, essentially a user would have to clear their session (ie restarting their browser), and then proceeding to the Backup Phrase step. On that first load, the decryptMnemonic step fails to properly overwrite the mnemonic in memory, it only overwrite the session store, thereby the next step would noramlly fail verification, since the mnemonic in memory, is a different one that gets generated on every browser session. It may be worth considering stopping this practice as well, if seedBlob is already present.