fix: unblur not working due to disabled
this should rectify text-blur toggle not working on certain platforms and browsers whilst the element is disabled. The readonly attribute should provide sufficient similariyt to disabled in that the text should not be editable, while still catching click events. May be worthwhile in adding a style for readonly attributes, to textareas, as it normally indicates something of importance to the user.