fix: tooltips on mobile
ironically by 'unfixing' their position
the fixed position worked great, in terms of using the entire screen
width, however, it came with a bad side-effect when the screen was
scrollable, in that they would be positioned to the element's initial
non-scrolled position, which doesn't look nice...

this commit should fix that, and also generally improves the animiations
between appearing and disappearing tooltips, previously it would
disappear more harshly, and was also timed, and just kind of awkward,
though functional and very simple, which. However, the above issue
required more attention by me, and here it is. Obviously more CSS had to
put in, and more edge case handling.

Also improved some of the tooltip contents